+++ packages/assets-controllers/src/TokenDetectionController.ts
@@ -4,10 +4,9 @@ import type {
   AccountsControllerSelectedEvmAccountChangeEvent,
 } from '@metamask/accounts-controller';
 import type {
-  RestrictedMessenger,
   ControllerGetStateAction,
   ControllerStateChangeEvent,
-} from '@metamask/base-controller';
+} from '@metamask/base-controller/next';
 import contractMap from '@metamask/contract-metadata';
 import {
   ASSET_TYPES,
@@ -22,6 +21,7 @@ import type {
   KeyringControllerLockEvent,
   KeyringControllerUnlockEvent,
 } from '@metamask/keyring-controller';
+import type { Messenger } from '@metamask/messenger';
 import type {
   NetworkClientId,
   NetworkControllerFindNetworkClientIdByChainIdAction,
@@ -158,12 +158,10 @@ export type AllowedEvents =
   | PreferencesControllerStateChangeEvent
   | TransactionControllerTransactionConfirmedEvent;
 
-export type TokenDetectionControllerMessenger = RestrictedMessenger<
+export type TokenDetectionControllerMessenger = Messenger<
   typeof controllerName,
   TokenDetectionControllerActions | AllowedActions,
-  TokenDetectionControllerEvents | AllowedEvents,
-  AllowedActions['type'],
-  AllowedEvents['type']
+  TokenDetectionControllerEvents | AllowedEvents
 >;
 
 /** The input to start polling for the {@link TokenDetectionController} */
@@ -275,7 +273,7 @@ export class TokenDetectionController extends StaticIntervalPollingController<To
    * Creates a TokenDetectionController instance.
    *
    * @param options - The controller options.
-   * @param options.messenger - The controller messaging system.
+   * @param options.messenger - The controller messenger.
    * @param options.disabled - If set to true, all network requests are blocked.
    * @param options.interval - Polling interval used to fetch new token rates
    * @param options.getBalancesInSingleCall - Gets the balances of a list of tokens for the given address.
@@ -321,7 +319,7 @@ export class TokenDetectionController extends StaticIntervalPollingController<To
       metadata: {},
     });
 
-    this.messagingSystem.registerActionHandler(
+    this.messenger.registerActionHandler(
       `${controllerName}:addDetectedTokensViaWs` as const,
       this.addDetectedTokensViaWs.bind(this),
     );
@@ -331,23 +329,22 @@ export class TokenDetectionController extends StaticIntervalPollingController<To
 
     this.#selectedAccountId = this.#getSelectedAccount().id;
 
-    const { tokensChainsCache } = this.messagingSystem.call(
+    const { tokensChainsCache } = this.messenger.call(
       'TokenListController:getState',
     );
 
     this.#tokensChainsCache = tokensChainsCache;
 
-    const { useTokenDetection: defaultUseTokenDetection } =
-      this.messagingSystem.call('PreferencesController:getState');
+    const { useTokenDetection: defaultUseTokenDetection } = this.messenger.call(
+      'PreferencesController:getState',
+    );
     this.#isDetectionEnabledFromPreferences = defaultUseTokenDetection;
 
     this.#getBalancesInSingleCall = getBalancesInSingleCall;
 
     this.#trackMetaMetricsEvent = trackMetaMetricsEvent;
 
-    const { isUnlocked } = this.messagingSystem.call(
-      'KeyringController:getState',
-    );
+    const { isUnlocked } = this.messenger.call('KeyringController:getState');
     this.#isUnlocked = isUnlocked;
 
     this.#accountsAPI.isAccountsAPIEnabled = useAccountsAPI;
@@ -359,20 +356,20 @@ export class TokenDetectionController extends StaticIntervalPollingController<To
   }
 
   /**
-   * Constructor helper for registering this controller's messaging system subscriptions to controller events.
+   * Constructor helper for registering this controller's messenger subscriptions to controller events.
    */
   #registerEventListeners() {
-    this.messagingSystem.subscribe('KeyringController:unlock', async () => {
+    this.messenger.subscribe('KeyringController:unlock', async () => {
       this.#isUnlocked = true;
       await this.#restartTokenDetection();
     });
 
-    this.messagingSystem.subscribe('KeyringController:lock', () => {
+    this.messenger.subscribe('KeyringController:lock', () => {
       this.#isUnlocked = false;
       this.#stopPolling();
     });
 
-    this.messagingSystem.subscribe(
+    this.messenger.subscribe(
       'TokenListController:stateChange',
       async ({ tokensChainsCache }) => {
         const isEqualValues = this.#compareTokensChainsCache(
@@ -385,7 +382,7 @@ export class TokenDetectionController extends StaticIntervalPollingController<To
       },
     );
 
-    this.messagingSystem.subscribe(
+    this.messenger.subscribe(
       'PreferencesController:stateChange',
       async ({ useTokenDetection }) => {
         const selectedAccount = this.#getSelectedAccount();
@@ -402,10 +399,10 @@ export class TokenDetectionController extends StaticIntervalPollingController<To
       },
     );
 
-    this.messagingSystem.subscribe(
+    this.messenger.subscribe(
       'AccountsController:selectedEvmAccountChange',
       async (selectedAccount) => {
-        const { networkConfigurationsByChainId } = this.messagingSystem.call(
+        const { networkConfigurationsByChainId } = this.messenger.call(
           'NetworkController:getState',
         );
 
@@ -422,7 +419,7 @@ export class TokenDetectionController extends StaticIntervalPollingController<To
       },
     );
 
-    this.messagingSystem.subscribe(
+    this.messenger.subscribe(
       'TransactionController:transactionConfirmed',
       async (transactionMeta) => {
         await this.detectTokens({
@@ -521,10 +518,10 @@ export class TokenDetectionController extends StaticIntervalPollingController<To
     chainIds: Hex[] | undefined,
   ): { chainId: Hex; networkClientId: NetworkClientId }[] {
     const { networkConfigurationsByChainId, selectedNetworkClientId } =
-      this.messagingSystem.call('NetworkController:getState');
+      this.messenger.call('NetworkController:getState');
 
     if (!chainIds) {
-      const networkConfiguration = this.messagingSystem.call(
+      const networkConfiguration = this.messenger.call(
         'NetworkController:getNetworkConfigurationByNetworkClientId',
         selectedNetworkClientId,
       );
@@ -647,7 +644,7 @@ export class TokenDetectionController extends StaticIntervalPollingController<To
     if (isMainnetDetectionInactive) {
       this.#tokensChainsCache = this.#getConvertedStaticMainnetTokenList();
     } else {
-      const { tokensChainsCache } = this.messagingSystem.call(
+      const { tokensChainsCache } = this.messenger.call(
         'TokenListController:getState',
       );
       this.#tokensChainsCache = tokensChainsCache ?? {};
@@ -747,7 +744,7 @@ export class TokenDetectionController extends StaticIntervalPollingController<To
     selectedAddress: string;
   }): string[][] {
     const { allTokens, allDetectedTokens, allIgnoredTokens } =
-      this.messagingSystem.call('TokensController:getState');
+      this.messenger.call('TokensController:getState');
     const [tokensAddresses, detectedTokensAddresses, ignoredTokensAddresses] = [
       allTokens,
       allDetectedTokens,
@@ -842,7 +839,7 @@ export class TokenDetectionController extends StaticIntervalPollingController<To
         const isTokenDetectionInactiveInMainnet =
           !this.#isDetectionEnabledFromPreferences &&
           chainId === ChainId.mainnet;
-        const { tokensChainsCache } = this.messagingSystem.call(
+        const { tokensChainsCache } = this.messenger.call(
           'TokenListController:getState',
         );
         this.#tokensChainsCache = isTokenDetectionInactiveInMainnet
@@ -883,12 +880,12 @@ export class TokenDetectionController extends StaticIntervalPollingController<To
             },
           });
 
-          const networkClientId = this.messagingSystem.call(
+          const networkClientId = this.messenger.call(
             'NetworkController:findNetworkClientIdByChainId',
             chainId,
           );
 
-          await this.messagingSystem.call(
+          await this.messenger.call(
             'TokensController:addTokens',
             tokensWithBalance,
             networkClientId,
@@ -1012,7 +1009,7 @@ export class TokenDetectionController extends StaticIntervalPollingController<To
           },
         });
 
-        await this.messagingSystem.call(
+        await this.messenger.call(
           'TokensController:addTokens',
           tokensWithBalance,
           networkClientId,
@@ -1086,12 +1083,12 @@ export class TokenDetectionController extends StaticIntervalPollingController<To
         },
       });
 
-      const networkClientId = this.messagingSystem.call(
+      const networkClientId = this.messenger.call(
         'NetworkController:findNetworkClientIdByChainId',
         chainId,
       );
 
-      await this.messagingSystem.call(
+      await this.messenger.call(
         'TokensController:addTokens',
         tokensWithBalance,
         networkClientId,
@@ -1100,12 +1097,12 @@ export class TokenDetectionController extends StaticIntervalPollingController<To
   }
 
   #getSelectedAccount() {
-    return this.messagingSystem.call('AccountsController:getSelectedAccount');
+    return this.messenger.call('AccountsController:getSelectedAccount');
   }
 
   #getSelectedAddress() {
     // If the address is not defined (or empty), we fallback to the currently selected account's address
-    const account = this.messagingSystem.call(
+    const account = this.messenger.call(
       'AccountsController:getAccount',
       this.#selectedAccountId,
     );