+++ packages/transaction-controller/src/TransactionController.ts
@@ -11,9 +11,9 @@ import type {
 import type {
   ControllerGetStateAction,
   ControllerStateChangeEvent,
-  RestrictedMessenger,
-} from '@metamask/base-controller';
-import { BaseController } from '@metamask/base-controller';
+  StateMetadata,
+} from '@metamask/base-controller/next';
+import { BaseController } from '@metamask/base-controller/next';
 import {
   query,
   ApprovalType,
@@ -27,6 +27,7 @@ import type {
   GasFeeState,
 } from '@metamask/gas-fee-controller';
 import type { KeyringControllerSignEip7702AuthorizationAction } from '@metamask/keyring-controller';
+import type { Messenger } from '@metamask/messenger';
 import type {
   BlockTracker,
   NetworkClientId,
@@ -177,35 +178,35 @@ import {
  * Metadata for the TransactionController state, describing how to "anonymize"
  * the state and which parts should be persisted.
  */
-const metadata = {
+const metadata: StateMetadata<TransactionControllerState> = {
   transactions: {
     includeInStateLogs: true,
     persist: true,
-    anonymous: false,
+    includeInDebugSnapshot: false,
     usedInUi: true,
   },
   transactionBatches: {
     includeInStateLogs: true,
     persist: true,
-    anonymous: false,
+    includeInDebugSnapshot: false,
     usedInUi: true,
   },
   methodData: {
     includeInStateLogs: true,
     persist: true,
-    anonymous: false,
+    includeInDebugSnapshot: false,
     usedInUi: true,
   },
   lastFetchedBlockNumbers: {
     includeInStateLogs: true,
     persist: true,
-    anonymous: false,
+    includeInDebugSnapshot: false,
     usedInUi: false,
   },
   submitHistory: {
     includeInStateLogs: true,
     persist: true,
-    anonymous: false,
+    includeInDebugSnapshot: false,
     usedInUi: false,
   },
 };
@@ -736,12 +737,10 @@ export type TransactionControllerEvents =
 /**
  * The messenger of the {@link TransactionController}.
  */
-export type TransactionControllerMessenger = RestrictedMessenger<
+export type TransactionControllerMessenger = Messenger<
   typeof controllerName,
   TransactionControllerActions | AllowedActions,
-  TransactionControllerEvents | AllowedEvents,
-  AllowedActions['type'],
-  AllowedEvents['type']
+  TransactionControllerEvents | AllowedEvents
 >;
 
 /**
@@ -933,7 +932,7 @@ export class TransactionController extends BaseController<
       },
     });
 
-    this.messagingSystem = messenger;
+    this.messenger = messenger;
 
     this.#afterAdd = hooks?.afterAdd ?? (() => Promise.resolve({}));
     this.#afterSign = hooks?.afterSign ?? (() => true);
@@ -981,7 +980,7 @@ export class TransactionController extends BaseController<
     this.#transactionHistoryLimit = transactionHistoryLimit;
 
     const findNetworkClientIdByChainId = (chainId: Hex) => {
-      return this.messagingSystem.call(
+      return this.messenger.call(
         `NetworkController:findNetworkClientIdByChainId`,
         chainId,
       );
@@ -990,7 +989,7 @@ export class TransactionController extends BaseController<
     this.#multichainTrackingHelper = new MultichainTrackingHelper({
       findNetworkClientIdByChainId,
       getNetworkClientById: ((networkClientId: NetworkClientId) => {
-        return this.messagingSystem.call(
+        return this.messenger.call(
           `NetworkController:getNetworkClientById`,
           networkClientId,
         );
@@ -1002,10 +1001,7 @@ export class TransactionController extends BaseController<
       createPendingTransactionTracker:
         this.#createPendingTransactionTracker.bind(this),
       onNetworkStateChange: (listener) => {
-        this.messagingSystem.subscribe(
-          'NetworkController:stateChange',
-          listener,
-        );
+        this.messenger.subscribe('NetworkController:stateChange', listener);
       },
     });
 
@@ -1021,12 +1017,9 @@ export class TransactionController extends BaseController<
       getTransactions: () => this.state.transactions,
       getTransactionBatches: () => this.state.transactionBatches,
       layer1GasFeeFlows: this.#layer1GasFeeFlows,
-      messenger: this.messagingSystem,
+      messenger: this.messenger,
       onStateChange: (listener) => {
-        this.messagingSystem.subscribe(
-          'TransactionController:stateChange',
-          listener,
-        );
+        this.messenger.subscribe('TransactionController:stateChange', listener);
       },
     });
 
@@ -1061,7 +1054,7 @@ export class TransactionController extends BaseController<
       includeTokenTransfers:
         this.#incomingTransactionOptions.includeTokenTransfers,
       isEnabled: this.#incomingTransactionOptions.isEnabled,
-      messenger: this.messagingSystem,
+      messenger: this.messenger,
       remoteTransactionSource: new AccountsApiRemoteTransactionSource(),
       trimTransactions: this.#trimTransactionsForState.bind(this),
       updateTransactions: this.#incomingTransactionOptions.updateTransactions,
@@ -1073,7 +1066,7 @@ export class TransactionController extends BaseController<
 
     // when transactionsController state changes
     // check for pending transactions and start polling if there are any
-    this.messagingSystem.subscribe(
+    this.messenger.subscribe(
       'TransactionController:stateChange',
       this.#checkForPendingTransactionAndStartPolling,
     );
@@ -1081,7 +1074,7 @@ export class TransactionController extends BaseController<
     new ResimulateHelper({
       simulateTransaction: this.#updateSimulationData.bind(this),
       onTransactionsUpdate: (listener) => {
-        this.messagingSystem.subscribe(
+        this.messenger.subscribe(
           'TransactionController:stateChange',
           listener,
           (controllerState) => controllerState.transactions,
@@ -1125,7 +1118,7 @@ export class TransactionController extends BaseController<
   async addTransactionBatch(
     request: TransactionBatchRequest,
   ): Promise<TransactionBatchResult> {
-    const { blockTracker } = this.messagingSystem.call(
+    const { blockTracker } = this.messenger.call(
       `NetworkController:getNetworkClientById`,
       request.networkClientId,
     );
@@ -1148,7 +1141,7 @@ export class TransactionController extends BaseController<
       getTransaction: (transactionId) =>
         this.#getTransactionOrThrow(transactionId),
       isSimulationEnabled: this.#isSimulationEnabled,
-      messenger: this.messagingSystem,
+      messenger: this.messenger,
       publishBatchHook: this.#publishBatchHook,
       publicKeyEIP7702: this.#publicKeyEIP7702,
       publishTransaction: (
@@ -1174,7 +1167,7 @@ export class TransactionController extends BaseController<
     return isAtomicBatchSupported({
       ...request,
       getEthQuery: (chainId) => this.#getEthQuery({ chainId }),
-      messenger: this.messagingSystem,
+      messenger: this.messenger,
       publicKeyEIP7702: this.#publicKeyEIP7702,
     });
   }
@@ -1362,7 +1355,7 @@ export class TransactionController extends BaseController<
         {
           isSwapsDisabled: this.#isSwapsDisabled,
           cancelTransaction: this.#rejectTransaction.bind(this),
-          messenger: this.messagingSystem,
+          messenger: this.messenger,
         },
       );
 
@@ -1394,7 +1387,7 @@ export class TransactionController extends BaseController<
         );
       }
 
-      this.messagingSystem.publish(
+      this.messenger.publish(
         `${controllerName}:unapprovedTransactionAdded`,
         addedTransactionMeta,
       );
@@ -1462,7 +1455,7 @@ export class TransactionController extends BaseController<
         txParams.value = '0x0';
       },
       afterSubmit: (newTransactionMeta) => {
-        this.messagingSystem.publish(
+        this.messenger.publish(
           `${controllerName}:transactionFinished`,
           newTransactionMeta,
         );
@@ -1501,7 +1494,7 @@ export class TransactionController extends BaseController<
       transactionId,
       transactionType: TransactionType.retry,
       afterSubmit: (newTransactionMeta) => {
-        this.messagingSystem.publish(
+        this.messenger.publish(
           `${controllerName}:speedupTransactionAdded`,
           newTransactionMeta,
         );
@@ -1617,12 +1610,12 @@ export class TransactionController extends BaseController<
     this.#addMetadata(newTransactionMeta);
 
     // speedUpTransaction has no approval request, so we assume the user has already approved the transaction
-    this.messagingSystem.publish(`${controllerName}:transactionApproved`, {
+    this.messenger.publish(`${controllerName}:transactionApproved`, {
       transactionMeta: newTransactionMeta,
       actionId,
     });
 
-    this.messagingSystem.publish(`${controllerName}:transactionSubmitted`, {
+    this.messenger.publish(`${controllerName}:transactionSubmitted`, {
       transactionMeta: newTransactionMeta,
       actionId,
     });
@@ -1658,7 +1651,7 @@ export class TransactionController extends BaseController<
       ignoreDelegationSignatures,
       isSimulationEnabled: this.#isSimulationEnabled(),
       getSimulationConfig: this.#getSimulationConfig,
-      messenger: this.messagingSystem,
+      messenger: this.messenger,
       txParams: transaction,
     });
 
@@ -1687,7 +1680,7 @@ export class TransactionController extends BaseController<
       ethQuery,
       isSimulationEnabled: this.#isSimulationEnabled(),
       getSimulationConfig: this.#getSimulationConfig,
-      messenger: this.messagingSystem,
+      messenger: this.messenger,
       txParams: transaction,
     });
 
@@ -1834,7 +1827,7 @@ export class TransactionController extends BaseController<
         throw error;
       });
 
-      this.messagingSystem.publish(
+      this.messenger.publish(
         `${controllerName}:transactionConfirmed`,
         updatedTransactionMeta,
       );
@@ -2178,7 +2171,7 @@ export class TransactionController extends BaseController<
 
     await updateTransactionLayer1GasFee({
       layer1GasFeeFlows: this.#layer1GasFeeFlows,
-      messenger: this.messagingSystem,
+      messenger: this.messenger,
       provider,
       transactionMeta: updatedTransaction,
     });
@@ -2372,7 +2365,7 @@ export class TransactionController extends BaseController<
         status as TransactionStatus,
       )
     ) {
-      this.messagingSystem.publish(
+      this.messenger.publish(
         `${controllerName}:transactionFinished`,
         updatedTransactionMeta,
       );
@@ -2506,7 +2499,7 @@ export class TransactionController extends BaseController<
     const gasFeeFlow = getGasFeeFlow(
       transactionMeta,
       this.#gasFeeFlows,
-      this.messagingSystem,
+      this.messenger,
     ) as GasFeeFlow;
 
     const ethQuery = new EthQuery(provider);
@@ -2518,7 +2511,7 @@ export class TransactionController extends BaseController<
     return gasFeeFlow.getGasFees({
       ethQuery,
       gasFeeControllerData,
-      messenger: this.messagingSystem,
+      messenger: this.messenger,
       transactionMeta,
     });
   }
@@ -2548,7 +2541,7 @@ export class TransactionController extends BaseController<
 
     return await getTransactionLayer1GasFee({
       layer1GasFeeFlows: this.#layer1GasFeeFlows,
-      messenger: this.messagingSystem,
+      messenger: this.messenger,
       provider,
       transactionMeta: {
         txParams: transactionParams,
@@ -2824,7 +2817,7 @@ export class TransactionController extends BaseController<
           gasFeeFlows: this.#gasFeeFlows,
           getGasFeeEstimates: this.#getGasFeeEstimates,
           getSavedGasFees: this.#getSavedGasFees.bind(this),
-          messenger: this.messagingSystem,
+          messenger: this.messenger,
           txMeta: transactionMeta,
         }),
     );
@@ -2834,7 +2827,7 @@ export class TransactionController extends BaseController<
       async () =>
         await updateTransactionLayer1GasFee({
           layer1GasFeeFlows: this.#layer1GasFeeFlows,
-          messenger: this.messagingSystem,
+          messenger: this.messenger,
           provider,
           transactionMeta,
         }),
@@ -2941,13 +2934,10 @@ export class TransactionController extends BaseController<
           const updatedTransactionMeta = this.#getTransaction(
             transactionId,
           ) as TransactionMeta;
-          this.messagingSystem.publish(
-            `${controllerName}:transactionApproved`,
-            {
-              transactionMeta: updatedTransactionMeta,
-              actionId,
-            },
-          );
+          this.messenger.publish(`${controllerName}:transactionApproved`, {
+            transactionMeta: updatedTransactionMeta,
+            actionId,
+          });
         }
       } catch (rawError: unknown) {
         const error = rawError as Error & { code?: number; data?: Json };
@@ -3079,7 +3069,7 @@ export class TransactionController extends BaseController<
 
       if (!(await this.#beforePublish(transactionMeta))) {
         log('Skipping publishing transaction based on hook');
-        this.messagingSystem.publish(
+        this.messenger.publish(
           `${controllerName}:transactionPublishingSkipped`,
           transactionMeta,
         );
@@ -3149,11 +3139,11 @@ export class TransactionController extends BaseController<
         },
       );
 
-      this.messagingSystem.publish(`${controllerName}:transactionSubmitted`, {
+      this.messenger.publish(`${controllerName}:transactionSubmitted`, {
         transactionMeta,
       });
 
-      this.messagingSystem.publish(
+      this.messenger.publish(
         `${controllerName}:transactionFinished`,
         transactionMeta,
       );
@@ -3211,7 +3201,7 @@ export class TransactionController extends BaseController<
       error: normalizeTxError(error ?? providerErrors.userRejectedRequest()),
     };
 
-    this.messagingSystem.publish(
+    this.messenger.publish(
       `${controllerName}:transactionFinished`,
       updatedTransactionMeta,
     );
@@ -3221,7 +3211,7 @@ export class TransactionController extends BaseController<
       updatedTransactionMeta,
     );
 
-    this.messagingSystem.publish(`${controllerName}:transactionRejected`, {
+    this.messenger.publish(`${controllerName}:transactionRejected`, {
       transactionMeta: updatedTransactionMeta,
       actionId,
     });
@@ -3323,7 +3313,7 @@ export class TransactionController extends BaseController<
       parentContext: traceContext,
     });
 
-    return (await this.messagingSystem.call(
+    return (await this.messenger.call(
       'ApprovalController:addRequest',
       {
         id,
@@ -3459,7 +3449,7 @@ export class TransactionController extends BaseController<
       );
     });
 
-    this.messagingSystem.publish(
+    this.messenger.publish(
       `${controllerName}:incomingTransactionsReceived`,
       finalTransactions,
     );
@@ -3608,7 +3598,7 @@ export class TransactionController extends BaseController<
       ...transactionMeta,
       status: TransactionStatus.dropped as const,
     };
-    this.messagingSystem.publish(`${controllerName}:transactionDropped`, {
+    this.messenger.publish(`${controllerName}:transactionDropped`, {
       transactionMeta: updatedTransactionMeta,
     });
     this.updateTransaction(
@@ -3698,7 +3688,7 @@ export class TransactionController extends BaseController<
 
     const signedAuthorizationList = await signAuthorizationList({
       authorizationList,
-      messenger: this.messagingSystem,
+      messenger: this.messenger,
       transactionMeta,
     });
 
@@ -3790,7 +3780,7 @@ export class TransactionController extends BaseController<
   }
 
   #onTransactionStatusChange(transactionMeta: TransactionMeta) {
-    this.messagingSystem.publish(`${controllerName}:transactionStatusUpdated`, {
+    this.messenger.publish(`${controllerName}:transactionStatusUpdated`, {
       transactionMeta,
     });
   }
@@ -3813,7 +3803,7 @@ export class TransactionController extends BaseController<
 
     this.#markNonceDuplicatesDropped(transactionMeta.id);
 
-    this.messagingSystem.publish(
+    this.messenger.publish(
       `${controllerName}:transactionConfirmed`,
       transactionMeta,
     );
@@ -3844,7 +3834,7 @@ export class TransactionController extends BaseController<
           updateTransaction: this.updateTransaction.bind(this),
         });
 
-      this.messagingSystem.publish(
+      this.messenger.publish(
         `${controllerName}:postTransactionBalanceUpdated`,
         {
           transactionMeta: updatedTransactionMeta,
@@ -3908,7 +3898,7 @@ export class TransactionController extends BaseController<
         this.#multichainTrackingHelper.acquireNonceLockForChainIdKey({
           chainId,
         }),
-      messenger: this.messagingSystem,
+      messenger: this.messenger,
       publishTransaction: (_ethQuery, transactionMeta) =>
         this.#publishTransaction(_ethQuery, transactionMeta, {
           skipSubmitHistory: true,
@@ -4184,7 +4174,7 @@ export class TransactionController extends BaseController<
         chainId,
         getSimulationConfig: this.#getSimulationConfig,
         isEIP7702GasFeeTokensEnabled: this.#isEIP7702GasFeeTokensEnabled,
-        messenger: this.messagingSystem,
+        messenger: this.messenger,
         publicKeyEIP7702: this.#publicKeyEIP7702,
         transactionMeta,
       });
@@ -4283,11 +4273,11 @@ export class TransactionController extends BaseController<
   }
 
   #getSelectedAccount() {
-    return this.messagingSystem.call('AccountsController:getSelectedAccount');
+    return this.messenger.call('AccountsController:getSelectedAccount');
   }
 
   #getInternalAccounts(): Hex[] {
-    const state = this.messagingSystem.call('AccountsController:getState');
+    const state = this.messenger.call('AccountsController:getState');
 
     return Object.values(state.internalAccounts?.accounts ?? {})
       .filter((account) => account.type === 'eip155:eoa')
@@ -4347,48 +4337,48 @@ export class TransactionController extends BaseController<
       isCustomNetwork,
       isSimulationEnabled: this.#isSimulationEnabled(),
       getSimulationConfig: this.#getSimulationConfig,
-      messenger: this.messagingSystem,
+      messenger: this.messenger,
       txMeta: transactionMeta,
     });
   }
 
   #registerActionHandlers(): void {
-    this.messagingSystem.registerActionHandler(
+    this.messenger.registerActionHandler(
       `${controllerName}:addTransaction`,
       this.addTransaction.bind(this),
     );
 
-    this.messagingSystem.registerActionHandler(
+    this.messenger.registerActionHandler(
       `${controllerName}:addTransactionBatch`,
       this.addTransactionBatch.bind(this),
     );
 
-    this.messagingSystem.registerActionHandler(
+    this.messenger.registerActionHandler(
       `${controllerName}:confirmExternalTransaction`,
       this.confirmExternalTransaction.bind(this),
     );
 
-    this.messagingSystem.registerActionHandler(
+    this.messenger.registerActionHandler(
       `${controllerName}:estimateGas`,
       this.estimateGas.bind(this),
     );
 
-    this.messagingSystem.registerActionHandler(
+    this.messenger.registerActionHandler(
       `${controllerName}:getNonceLock`,
       this.getNonceLock.bind(this),
     );
 
-    this.messagingSystem.registerActionHandler(
+    this.messenger.registerActionHandler(
       `${controllerName}:getTransactions`,
       this.getTransactions.bind(this),
     );
 
-    this.messagingSystem.registerActionHandler(
+    this.messenger.registerActionHandler(
       `${controllerName}:updateCustodialTransaction`,
       this.updateCustodialTransaction.bind(this),
     );
 
-    this.messagingSystem.registerActionHandler(
+    this.messenger.registerActionHandler(
       `${controllerName}:updateTransaction`,
       this.updateTransaction.bind(this),
     );
@@ -4462,7 +4452,7 @@ export class TransactionController extends BaseController<
       };
     }
 
-    this.messagingSystem.publish(`${controllerName}:transactionFailed`, {
+    this.messenger.publish(`${controllerName}:transactionFailed`, {
       actionId,
       error: error.message,
       transactionMeta: newTransactionMeta,
@@ -4470,7 +4460,7 @@ export class TransactionController extends BaseController<
 
     this.#onTransactionStatusChange(newTransactionMeta);
 
-    this.messagingSystem.publish(
+    this.messenger.publish(
       `${controllerName}:transactionFinished`,
       newTransactionMeta,
     );